Fixed issue where some param client calls failed for remote EK80 requests
A number of the "get" methods of EK80_REST_Client failed to pass a configured instance of ApiClient to the underlying Api classes so the host would default to "localhost" which results in failed/refused connections when interacting with an EK80 server on the LAN. Also added a workaround to handle specifying the EK80 address as "localhost" or "127.0.0.1". http acls created by EK80 only permit connections to "localhost" but endpoint creation fails when provided "localhost" so endpoints for local access are always created with 127.0.0.1.